backend/requirements.txt
fastapi==0.116.1
uvicorn==0.30.6
pydantic==2.10.1
requests==2.32.3
python-dotenv==1.0.1
python-multipart==0.0.12

